rename CommentService into CommentManager + rename method intoEmbedLinkYoutbe into intoEmbedLinkYoutube
