+++ src/Service/LinkYoutubeService.php
@@ -8,7 +8,7 @@ class LinkYoutubeService
     /**
      * function to convert shareLink into displayLink of Youtube
      */
-    public function intoEmbedLinkYoutbe(string $shareLink): string
+    public function intoEmbedLinkYoutube(string $shareLink): string
     {
         $displayLink = str_replace('youtu.be', 'youtube.com/embed', $shareLink);
